+++ plugin/telegram/update.go
@@ -1,7 +1,7 @@
 package telegram
 
 type Update struct {
-	UpdateID      int            `json:"update_id"`
+	UpdateID      int64          `json:"update_id"`
 	Message       *Message       `json:"message"`
 	CallbackQuery *CallbackQuery `json:"callback_query"`
 }
